refactor(native): extract sidecar discovery
